requirements.txt
firebase
pymongo
httpx
google
fastapi
PySide6
python-dotenv
pyinstaller
google-cloud-firestore
Kivy
Kivymd
matplotlib
starlette
uvicorn
pydantic
regex